modify #63 ファイルサイズが小さくなったので認証と認可のメソッドをauth.tsに集約
