+++ serverless.yml
@@ -68,14 +68,14 @@ functions:
           authorizer: ${self:custom.authorizer}
           cors: true
   authentication:
-    handler: authentication.authentication
+    handler: auth.authentication
     events:
       - http:
           path: auth/authentication
           method: post
           cors: true
   authorization:
-    handler: authorization.authorization
+    handler: auth.authorization
 
 resources:
   Resources:
